+++ src/main/scala/PspList.scala
@@ -4,7 +4,7 @@ package core
 object PspList {
   implicit def newBuilder[A] : PspCanBuild[A, PspList[A]] = PspCanBuild(fromForeach)
 
-  def to(start: Int, end: Int): PspList[Int] = fromForeach(Foreach.to(start, end))
+  def to(start: Long, end: Long): PspList[Long] = fromForeach(Foreach.to(start, end))
 
   // implicit def nilIsCovariant[A](xs: nil.type): PspList[A] = empty[A]
   def fromForeach[A](xs: Foreach[A]): PspList[A] = xs match {
